Create new hook : useConfiguration
