+++ src/components/app/app.js
@@ -1,30 +1,13 @@
-import React, { useState, useEffect } from 'react';
+import React from 'react';
 import root from 'react-shadow/styled-components';
 import Rooter from 'components/router';
 import D from 'i18n';
 import Preloader from 'components/shared/preloader';
-import { QUEEN_URL } from 'utils/constants';
+import { useConfiguration } from 'utils/hook';
+import customStyle from './app.style';
 
 const App = () => {
-  const customStyle = {
-    margin: 'auto',
-    height: '100vh',
-    fontFamily: "'Gotham SSm A', 'Gotham SSm B', sans-serif",
-    backgroundColor: '#c3ddff',
-  };
-  const [configuration, setConfiguration] = useState(null);
-  useEffect(() => {
-    if (!configuration) {
-      const loadConfiguration = async () => {
-        const response = await fetch(`${QUEEN_URL}/configuration.json`);
-        const configurationResponse = await response.json();
-        configurationResponse.standalone =
-          configurationResponse.QUEEN_URL === window.location.origin;
-        setConfiguration(configurationResponse);
-      };
-      loadConfiguration();
-    }
-  }, [configuration]);
+  const { configuration } = useConfiguration();
 
   return (
     <root.div id="queen-container" style={customStyle}>
