+++ src/components/app/app.style.js
@@ -1,0 +1,6 @@
+export default {
+  margin: 'auto',
+  height: '100vh',
+  fontFamily: "'Gotham SSm A', 'Gotham SSm B', sans-serif",
+  backgroundColor: '#c3ddff',
+};
