+++ src/utils/hook/index.js
@@ -1,1 +1,2 @@
 export * from './auth';
+export * from './configuration';
